+++ packages/adherev/changelog/CHANGELOG.md
@@ -1,3 +1,16 @@
+# 1.0.13
+
+***
+
+2022-11-22
+
+* **adhere-ui-historyback**
+  - 修改为返回的判断逻辑
+* **adhere-util**
+  - color加入新方法
+* **adhere-ui-writingboard**
+  - 加入导出图片设置背景色
+
 # 1.0.12
 
 ***